+++ tests/ttnn/nightly/unit_tests/operations/pool/test_avgpool2d.py
@@ -15,6 +15,11 @@ from models.demos.deepseek_v3.utils.config_helpers import (
 )
 from models.common.utility_functions import is_blackhole
 from tests.ttnn.nightly.unit_tests.operations.pool.test_maxpool2d import HS
+from tests.sweep_framework.sweep_utils.pool2d_common import (
+    randomize_tensor,
+    parse_padding,
+    compute_output_shape,
+)
 
 SliceWidth = ttnn.Op2DDRAMSliceWidth
 SliceHeight = ttnn.Op2DDRAMSliceHeight
@@ -81,16 +86,6 @@ def tensor_map():
     return tensor_map
 
 
-def randomize_tensor(tensor_map, tensor_shape):
-    tensor_shape = tuple(tensor_shape)
-    if tensor_shape in tensor_map.keys():
-        torch_tensor = tensor_map[tensor_shape]
-    else:
-        torch_tensor = torch.randn(tensor_shape, dtype=torch.bfloat16)
-        tensor_map[tensor_shape] = torch_tensor
-    return torch_tensor
-
-
 def run_avg_pool2d(
     device,
     tensor_map,
@@ -117,20 +112,7 @@ def run_avg_pool2d(
     stride_h, stride_w = stride
     dilation_h = dilation_w = 1  # avg pool does not yet support dilation
 
-    # handle both 2D and 4D padding
-    padding_is_4d = False
-    if len(padding) == 2:
-        pad_h = int(padding[0] * 2)
-        pad_w = int(padding[1] * 2)
-        pad_t = pad_b = padding[0]
-        pad_l = pad_r = padding[1]
-    elif len(padding) == 4:
-        padding_is_4d = True
-        pad_t, pad_b, pad_l, pad_r = padding
-        pad_h = pad_t + pad_b
-        pad_w = pad_l + pad_r
-    else:
-        raise ValueError(f"Padding must be 2D or 4D tuple, got {len(padding)}D")
+    pad_t, pad_b, pad_l, pad_r, pad_h, pad_w, padding_is_4d = parse_padding(padding)
 
     if (out_dtype == ttnn.bfloat8_b or out_dtype == ttnn.bfloat4_b) and output_layout == ttnn.ROW_MAJOR_LAYOUT:
         pytest.skip("BFLOAT8_B/BFLOAT4_B output data format is not supported with ROW_MAJOR layout")
@@ -160,19 +142,21 @@ def run_avg_pool2d(
     if (in_h + pad_h) < kernel_h or (in_w + pad_w) < kernel_w:
         pytest.skip("kernel is too large for the padded tensor")
 
-    ceil_mode_out_shape_adj = False
-    if ceil_mode:
-        out_h = math.ceil((in_h + pad_h - dilation_h * (kernel_h - 1) - 1) / stride_h) + 1
-        out_w = math.ceil((in_w + pad_w - dilation_w * (kernel_w - 1) - 1) / stride_w) + 1
-        if ((out_h - 1) * stride_h) >= (in_h + pad_t):
-            ceil_mode_out_shape_adj = True
-            out_h -= 1
-        if ((out_w - 1) * stride_w) >= (in_w + pad_l):
-            ceil_mode_out_shape_adj = True
-            out_w -= 1
-    else:
-        out_h = math.floor((in_h + pad_h - dilation_h * (kernel_h - 1) - 1) / stride_h) + 1
-        out_w = math.floor((in_w + pad_w - dilation_w * (kernel_w - 1) - 1) / stride_w) + 1
+    out_h, out_w, ceil_mode_out_shape_adj = compute_output_shape(
+        in_h,
+        in_w,
+        kernel_h,
+        kernel_w,
+        stride_h,
+        stride_w,
+        dilation_h,
+        dilation_w,
+        pad_h,
+        pad_w,
+        pad_t,
+        pad_l,
+        ceil_mode,
+    )
 
     # using non-zero seed to avoid random spike in floating point error on single element of the
     # 1x256x56x56 tensor with divisor_override=5 and 5x5 kernel resulting in rtol=0.015 for that element
