+++ tests/ttnn/nightly/unit_tests/operations/pool/test_max_pool2d_sweeps_mpwi.py
@@ -2,7 +2,7 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-from tests.sweep_framework.sweep_utils.max_pool2d_with_indices_common import run_max_pool2d_with_indices
+from tests.sweep_framework.sweep_utils.pool2d_common import run_max_pool2d_with_indices
 
 import pytest
 import ttnn
